+++ libs/cmcd/src/isCmcdEventKey.ts
@@ -1,5 +1,6 @@
 import { CMCD_COMMON_KEYS } from './CMCD_COMMON_KEYS.ts'
 import { CMCD_EVENT_KEYS } from './CMCD_EVENT_KEYS.ts'
+import { CMCD_RESPONSE_KEYS } from './CMCD_RESPONSE_KEYS.js'
 import type { CmcdEvent } from './CmcdEvent.ts'
 import { isCmcdCustomKey } from './isCmcdCustomKey.ts'
 
@@ -19,5 +20,6 @@ import { isCmcdCustomKey } from './isCmcdCustomKey.ts'
 export function isCmcdEventKey(key: string): key is keyof CmcdEvent {
 	return CMCD_COMMON_KEYS.includes(key as any) ||
 		CMCD_EVENT_KEYS.includes(key as any) ||
+		CMCD_RESPONSE_KEYS.includes(key as any) ||
 		isCmcdCustomKey(key as any)
 }