+++ libs/cmcd/test/data/QUERY_TARGET.ts
@@ -3,6 +3,6 @@ import { CmcdReportingMode, CmcdTransmissionMode, type CmcdReportTarget } from '
 export const QUERY_TARGET: CmcdReportTarget = {
 	url: 'https://hello.world',
 	version: 2,
-	reportingMode: CmcdReportingMode.RESPONSE,
+	reportingMode: CmcdReportingMode.EVENT,
 	transmissionMode: CmcdTransmissionMode.QUERY,
 }