+++ libs/cmcd/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.0.1] - 2025-12-1
+
+### Changed
+
+- Removed `response mode` from CMCD encoding
+- Added `response received` event and filtering
+
 ## [1.0.0] - 2025-10-24
 
 ### Changed